Implement Output File Creation
We don't need to bail if the output file name is null, because we
can provide a useful default.  However, if we can't even create
the output file, we should bail, because there's little point in
generating a log of changes from which you'd manually have to
create the output file.

Additionally, we also want to make sure that if an output file
exists already, we remove it so whatever we're currently working
on is the only thing that shows up in the output file.