Wire Everything Together
We finally implement the main method to tie the entire program
together.  We first check args so we don't waste time on bogus
input, and then check that we can create/access the output file,
again so we don't waste time doing the main body of work for
nothing.

If we have an issue on final write, we terminate with an abnormal
status code, but the log of changes should still be available since
the dedupe work is done.